pyproject.toml
[tool.poetry]
name = "remindo_bot"
version = "0.1.0"
description = "Made for the forgetful, Remindo will keep you on track with what you need to get done - even in a distracting platform like Discord"
authors = ["ChemDev"]

[tool.poetry.dependencies]
python = "3.9.2"
"discord.py" = "^1.7.3"
pip = "^22.2.2"
install = "^1.3.5"
ez_setup = "^0.9"
pydantic = "^1.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
